Initial commit
Added basic parser, does not do much at the moment, along with skeleton of
program, test file, LICENSE, et al.

This program is meant to produce C code for serialization/deserialization of
CAN messages given a DBC file, perhaps eventually it will be capable of
producing other output as well.

For the moment the goal is to get it reading a full DBC file (not just parts of
it) and then spitting out C code that can deal with it.