Generate untested C code, fixed bug in y_mx_c
The program can now generate C code from a DBC file, although floats are
unhandled, and there is still a lot of work to do (scalings and offsets,
minimums and maximums, fault and signal not available values (which still need
to be extracted from the DBC file) are just some of the areas that need
fixing). The code it generates is untested, so it could largely be garbage, but
it is a start.