Added XSD, Fixed XML generation, Printing out grammar, ...
An XSD file has been added which can be used to validate the generated XML.

The grammar can now be printed out.

A minor bug in the XML generation has been fixed.

A file has been added for specify the DBC file format, but it has not been
populated (dbc.md).