Created 'god object', formatting changes, removed unneeded code
A single object which contains all known types for this CAN interface
is now generated.

The order in which things are generated has been changed as well.

Some unneeded code that was generated has been removed, this was in the
encoding/decode functions.

There have been some minor formatting changes.