See issue #31, added -D option
This commit adds a '-D' option which makes the C code generated use only
the 'double' type for the encoded and decoded signals. This should have
enough range for *most* uses, and anyway is optional. As the encode and
decode functions now have the same prototype it would be possible to
make a single encode/decode function, like has been done with
pack/unpack...this would make the library much easier to use.